Good optimization code.

+++ SacredUtils/resources/bin/CheckAvailabilityUpdateTemp.cs
@@ -8,20 +8,11 @@ namespace SacredUtils.resources.bin
         {
             if (AppSettings.ApplicationSettings.RemoveTempContent)
             {
-                if (File.Exists("mnxupdater.exe"))
-                {
-                    File.Delete("mnxupdater.exe");
-                }
+                if (File.Exists("mnxupdater.exe")) { File.Delete("mnxupdater.exe"); }
 
-                if (File.Exists("_newVersionSacredUtilsTemp.exe"))
-                {
-                    File.Delete("_newVersionSacredUtilsTemp.exe");
-                }
+                if (File.Exists("_newVersionSacredUtilsTemp.exe")) { File.Delete("_newVersionSacredUtilsTemp.exe"); }
 
-                if (File.Exists("updater-crash-report.log"))
-                {
-                    File.Delete("updater-crash-report.log");
-                }
+                if (File.Exists("updater-crash-report.log")) { File.Delete("updater-crash-report.log"); }
             }
         }
     }
